+++ mpogemp/GEDROIDS.C
@@ -201,7 +201,7 @@ for (zothusn=0; zothusn<nterms; zothusn++)
 		ddist *= 10000;
 		if (ddist > shipclass[wptr->shpclass].scanrange)
 			{
-			bearing = (int)(cbearing(&wptr->coord,&ptr->coord,wptr->heading));
+			bearing = cbearing(&wptr->coord,&ptr->coord,wptr->heading);
 			prfmsg(DROIDNEW,bearing);
 			outprfge(FILTER,zothusn);
 			}
@@ -1071,7 +1071,7 @@ WARSHP	*wptr;
 {
 if (!neutral(&ptr->coord) && !neutral(&wptr->coord) && wptr->cloak != 10 && gernd()%(4-(shipclass[ptr->shpclass].tough_factor/2)) == 0)
 	{
-	ptr->degrees = (int)(cbearing(&ptr->coord,&wptr->coord,ptr->heading));
+	ptr->degrees = cbearing(&ptr->coord,&wptr->coord,ptr->heading);
 	if (wptr->where == 1 && ptr->where == 1)
 		firehp(ptr,usrn);
 	else
